fixes on remove, style updates
